requirements.txt
# Drone AI Command System - NVIDIA GTC 2026
# Core dependencies for natural language drone control

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LLM and AI frameworks
langchain>=0.1.0
langgraph>=0.0.40

# HTTP clients for API calls
httpx>=0.25.0
aiohttp>=3.9.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# CLI and output formatting
rich>=13.0.0

# Optional: for Isaac Sim integration
# omni-isaac-sim>=2023.1.0
